move itemtypes to own file, move core into init test
